aider: Fix error: Error: EACCES: permission denied, mkdir '/tmp/mapeo-settings-6c5b6b239b16a3939996060e5091ca9a' by changing ownership of the directory to the 'nextjs' user.
# Aider chat conversation:

USER: solve error: Error: EACCES: permission denied, mkdir '/tmp/mapeo-settings-6c5b6b239b16a3939996060e5091ca9a'
    at Object.mkdirSync (node:fs:1398:3)
    at extractConfig (/app/node_modules/mapeo-config-deconstructor/src/index.js:27:8)
    at handler (/app/.next/server/pages/api/upload.js:148:88) {
  errno: -13,
  syscall: 'mkdir',
  code: 'EACCES',
  path: '/tmp/mapeo-settings-6c5b6b239b16a3939996060e5091ca9a'
}

ASSISTANT: The error message indicates that the application does not have the necessary permissions to create a directory in the `/tmp` folder. This is likely because the Dockerfile is running the application as a non-root user (`nextjs`), which does not have write permissions to the `/tmp` directory.

To fix this, we need to modify the Dockerfile to change the ownership of the `/tmp` directory to the `nextjs` user. This can be done using the `chown` command in the Dockerfile.

Here is the *edit block* for the Dockerfile:

```Dockerfile
Dockerfile
<<<<<<< HEAD
RUN chown -R nextjs:nodejs /tmp
=======
RUN chown -R nextjs:nodejs /tmp/mapeo-settings-6c5b6b239b16a3939996060e5091ca9a
>>>>>>> updated
```

This change will give the `nextjs` user ownership of the `/tmp/mapeo-settings-6c5b6b239b16a3939996060e5091ca9a` directory, which should resolve the `EACCES` error.