aider: Added `delete` and `post` methods to handle deleting and creating JSON files respectively.
# Aider chat conversation:

USER: Please generate code for two methods: a `delete` method and a `post` method. The `delete` method should search within the `presetsDir` directory and locate the file named with the `slug.json`, then proceed to delete it. The `post` method, on the other hand, should take the JSON data from the request body and create a new JSON file. The filename of the new JSON file should be the slugified version of a specified name.

ASSISTANT: To implement the `delete` and `post` methods, we will modify the `pages/api/project/[...id].ts` file.

The `delete` method will be added to the existing conditional statements that check the request method. It will use the `fs.unlinkSync` function to delete the file named `slug.json` from the `presetsDir` directory.

The `post` method will also be added to the conditional statements. It will take the JSON data from the request body, create a new JSON file in the `presetsDir` directory, and write the data to this file. The filename will be the slugified version of a specified name.

Here are the changes:

```typescript
pages/api/project/[...id].ts
<<<<<<< HEAD
		} else if (req.method === "PUT") {
			const { slug, name, color, sort, icon, projectName } = req.body;
			if (!slug) throw Error('No slug passed')
			console.log('Received request body:', req.body);
			const presetPath = `${presetsDir}/${slug}.json`;
			console.log('Preset path:', presetPath);
			let presetData: PresetData = {
				name,
				color,
				sort: parseInt(sort) || 0,
				icon,
			};
			console.log('Preset data:', presetData);
			if (fs.existsSync(presetPath)) {
				const existingData = await JSON.parse(fs.readFileSync(presetPath, 'utf-8'));
				presetData = { ...existingData, ...presetData };
			}
			await fs.writeFileSync(presetPath, JSON.stringify(presetData));
			console.log('Preset data written to file');
			const updatedData = await fetchPresets(id, presetsDir)
			res.status(200).json({ data: updatedData, error: null });
			console.log('Response sent');
		}
=======
		} else if (req.method === "PUT") {
			// ... existing PUT method code ...
		} else if (req.method === "DELETE") {
			const { slug } = req.body;
			if (!slug) throw Error('No slug passed');
			const presetPath = `${presetsDir}/${slug}.json`;
			if (fs.existsSync(presetPath)) {
				fs.unlinkSync(presetPath);
				console.log('Preset file deleted');
			} else {
				throw Error('Preset file not found');
			}
			const updatedData = await fetchPresets(id, presetsDir)
			res.status(200).json({ data: updatedData, error: null });
			console.log('Response sent');
		} else if (req.method === "POST") {
			const { slug, name, color, sort, icon } = req.body;
			if (!slug) throw Error('No slug passed');
			const presetPath = `${presetsDir}/${slug}.json`;
			let presetData: PresetData = {
				name,
				color,
				sort: parseInt(sort) || 0,
				icon,
			};
			await fs.writeFileSync(presetPath, JSON.stringify(presetData));
			console.log('Preset data written to file');
			const updatedData = await fetchPresets(id, presetsDir)
			res.status(200).json({ data: updatedData, error: null });
			console.log('Response sent');
		}
>>>>>>> updated
```

Please note that the `slug` is expected to be passed in the request body for both the `delete` and `post` methods. If it's not provided, an error will be thrown. Also, the `post` method will overwrite any existing file with the same name. If you want to prevent this, you can add a check to see if the file already exists and throw an error if it does.